+++ src/Microsoft.IdentityModel.Tokens.Saml/Properties/AssemblyInfo.cs
@@ -20,14 +20,14 @@ using System;
 using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Microsoft.IdentityModel.Protocols")]
-[assembly: AssemblyDescription("Contains classes for OpenId Connect protocol.")]
+[assembly: AssemblyTitle("Microsoft.IdentityModel.Tokens.Saml")]
+[assembly: AssemblyDescription("Contains classes for Saml1 and Saml2 token processing.")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyProduct("Microsoft.IdentityModel.Protocols")]
+[assembly: AssemblyProduct("Microsoft.IdentityModel.Tokens.Saml")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(true)]
-[assembly: Guid("DDA17899-06A2-4DDF-A552-30A5EBF0E2B0")]
+[assembly: Guid("63F6E63C-300D-4D4F-9364-0087E4F18477")]
 [assembly: AssemblyMetadata("Serviceable", "True")]
-[assembly: AssemblyFileVersion("2.0.0.0")]
+[assembly: AssemblyFileVersion("5.0.0.0")]
