+++ src/Microsoft.IdentityModel.Tokens.Saml/SamlAudienceRestrictionCondition.cs
@@ -1,6 +1,10 @@
-﻿namespace Microsoft.IdentityModel.Tokens.Saml
+﻿using System;
+using System.Collections.Generic;
+
+namespace Microsoft.IdentityModel.Tokens.Saml
 {
     public class SamlAudienceRestrictionCondition : SamlCondition
     {
+        public IList<Uri> Audiences { get; set; }
     }
 }
