+++ src/Microsoft.IdentityModel.Tokens/Base64UrlEncoder.cs
