+++ src/Microsoft.IdentityModel.Tokens/Properties/AssemblyInfo.cs
@@ -20,14 +20,14 @@ using System;
 using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Microsoft.IdentityModel.Protocols")]
-[assembly: AssemblyDescription("Contains classes for OpenId Connect protocol.")]
+[assembly: AssemblyTitle("Microsoft.IdentityModel.Tokens")]
+[assembly: AssemblyDescription("Contains classes SecurityToken and SecurityKey processing.")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyProduct("Microsoft.IdentityModel.Protocols")]
+[assembly: AssemblyProduct("Microsoft.IdentityModel.Tokens")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(true)]
-[assembly: Guid("DDA17899-06A2-4DDF-A552-30A5EBF0E2B0")]
+[assembly: Guid("90D6739C-7EBB-4A52-B4C6-B02BC726A894")]
 [assembly: AssemblyMetadata("Serviceable", "True")]
-[assembly: AssemblyFileVersion("2.0.0.0")]
+[assembly: AssemblyFileVersion("5.0.0.0")]
