+++ src/Microsoft.IdentityModel.Tokens/SecurityTokenDescriptor.cs
@@ -24,11 +24,12 @@ namespace Microsoft.IdentityModel.Tokens
 {
     public class SecurityTokenDescriptor
     {
-        public DateTime? NotBefore { get; set; }
-        public DateTime? Expires { get; set; }
-        public SigningCredentials SigningCredentials { get; set; }
-        public IEnumerable<Claim> Claims { get; set; }
-        public String Issuer { get; set; }
         public String Audience { get; set; }
+        public IEnumerable<Claim> Claims { get; set; }
+        public DateTime? Expires { get; set; }
+        public DateTime? IssuedAt { get; set; }
+        public String Issuer { get; set; }
+        public DateTime? NotBefore { get; set; }
+        public SigningCredentials SigningCredentials { get; set; }
     }
 }
