+++ test/Microsoft.IdentityModel.Logging.Tests/LoggerTests.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Diagnostics.Tracing;
+using Microsoft.IdentityModel.Tokens;
+using Microsoft.IdentityModel.Tokens.Jwt;
 using Xunit;
 
 namespace Microsoft.IdentityModel.Logging.Tests
