+++ test/Microsoft.IdentityModel.Protocols.Tests/project.json
@@ -1,23 +1,27 @@
 {
-    "version": "2.0.0-*",
-    "dependencies": {
-      "Microsoft.IdentityModel.Protocols": "2.0.0-*",
-      "Microsoft.IdentityModel.Tokens": "5.0.0-*",
-      "Microsoft.IdentityModel.Tokens.Jwt": "5.0.0-*",
-      "Microsoft.IdentityModel.Tokens.Saml" : "5.0.0-*",
-      "Newtonsoft.Json": "6.0.6",
-      "xunit.runner.aspnet": "2.0.0-aspnet-*"
-    },
-    "compile": ["**\\*.cs", "..\\Shared\\**\\*.cs"],
-    "commands": {
-        "test": "xunit.runner.aspnet"
-    },
-    "frameworks": {
-        "dnx451": {
-        }
-    },
-    "compilationOptions": {
-        "allowUnsafe": false,
-        "warningsAsErrors": true
+  "version": "2.0.0-*",
+  "dependencies": {
+    "Microsoft.IdentityModel.Protocols": "2.0.0-*",
+    "Microsoft.IdentityModel.Tokens": "5.0.0-*",
+    "Microsoft.IdentityModel.Tokens.Jwt": "5.0.0-*",
+    "Microsoft.IdentityModel.Tokens.Saml": "5.0.0-*",
+    "Newtonsoft.Json": "6.0.6",
+    "xunit.runner.aspnet": "2.0.0-aspnet-*"
+  },
+  "exclude": [
+    "CrossTokenTests.cs",
+    "WsFederationMessageTests.cs"
+  ],
+  "compile": ["**\\*.cs", "..\\Shared\\**\\*.cs"],
+  "commands": {
+    "test": "xunit.runner.aspnet"
+  },
+  "frameworks": {
+    "dnx451": {
     }
+  },
+  "compilationOptions": {
+    "allowUnsafe": false,
+    "warningsAsErrors": true
+  }
 }
