+++ test/Microsoft.IdentityModel.Tokens.Jwt.Tests/AudienceValidation.cs
@@ -18,6 +18,7 @@
 
 using System.Collections.Generic;
 using System.Security.Claims;
+using Microsoft.IdentityModel.Test;
 using Xunit;
 
 namespace Microsoft.IdentityModel.Tokens.Jwt.Test
