+++ test/Microsoft.IdentityModel.Tokens.Jwt.Tests/JwtDerivedTestTypes.cs
@@ -159,213 +159,4 @@ namespace Microsoft.IdentityModel.Tokens.Jwt.Test
         }
     }
 
-    public class NotAsymmetricOrSymmetricSecurityKey : SecurityKey
-    {
-        public override SignatureProvider GetSignatureProvider(string algorithm, bool verifyOnly)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override int KeySize
-        {
-            get { throw new NotImplementedException(); }
-        }
-
-        public static SecurityKey New { get { return new NotAsymmetricOrSymmetricSecurityKey(); } }
-    }
-
-    public class ReturnNullSymmetricSecurityKey : SymmetricSecurityKey
-    {
-        public ReturnNullSymmetricSecurityKey(byte[] keyBytes)
-            : base(keyBytes)
-        { }
-
-        public SymmetricSecurityKey SymmetricSecurityKey { get; set; }
-
-        public override byte[] Key
-        {
-            get
-            {
-                if (SymmetricSecurityKey == null)
-                {
-                    return null;
-                }
-
-                return SymmetricSecurityKey.Key;
-            }
-        }
-
-        public override int KeySize
-        {
-            get
-            {
-                if (SymmetricSecurityKey != null)
-                {
-                    return SymmetricSecurityKey.KeySize;
-                }
-
-                return 256;
-            }
-        }
-    }
-
-    public class ReturnNullAsymmetricSecurityKey : AsymmetricSecurityKey
-    {
-        public ReturnNullAsymmetricSecurityKey() { }
-
-        public override bool HasPrivateKey
-        {
-            get { throw new NotImplementedException(); }
-        }
-
-        public override SignatureProvider GetSignatureProvider(string algorithm, bool verifyOnly)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override int KeySize
-        {
-            get { throw new NotImplementedException(); }
-        }
-
-        public override bool HasPublicKey
-        {
-            get
-            {
-                throw new NotImplementedException();
-            }
-        }
-    }
-
-    /// <summary>
-    /// Useful for trigging an exception.
-    /// </summary>
-    public class FaultingAsymmetricSecurityKey : AsymmetricSecurityKey
-    {
-        AsymmetricSecurityKey _key;
-
-        public FaultingAsymmetricSecurityKey(AsymmetricSecurityKey key = null, AsymmetricAlgorithm agorithm = null, bool hasPrivateKey = false)
-        {
-            _key = key;
-        }
-
-        public override bool HasPrivateKey
-        {
-            get { return _key.HasPrivateKey; }
-        }
-
-        public override bool HasPublicKey
-        {
-            get
-            {
-                throw new NotImplementedException();
-            }
-        }
-
-        public override int KeySize { get { return _key.KeySize; } }
-
-        public override SignatureProvider GetSignatureProvider(string algorithm, bool verifyOnly)
-        {
-            throw new NotImplementedException();
-        }
-
-    }
-
-    /// <summary>
-    /// This works that is a parameter is null, we throw the exception when asked for the property
-    /// </summary>
-    public class FaultingSymmetricSecurityKey : SymmetricSecurityKey
-    {
-        Exception _throwMe;
-        KeyedHashAlgorithm _keyedHash;
-        SymmetricSecurityKey _key;
-        SymmetricAlgorithm _agorithm;
-        byte[] _keyBytes;
-
-        public FaultingSymmetricSecurityKey(SymmetricSecurityKey key, Exception throwMe, SymmetricAlgorithm algorithm = null, KeyedHashAlgorithm keyedHash = null, byte[] keyBytes = null)
-            : base(keyBytes)
-        {
-            _throwMe = throwMe;
-            _key = key;
-            _keyedHash = keyedHash;
-            _agorithm = algorithm;
-            _keyBytes = keyBytes;
-        }
-
-        public override byte[] Key
-        {
-            get
-            {
-                if (_throwMe != null)
-                    throw _throwMe;
-
-                return _keyBytes;
-            }
-        }
-
-        public override int KeySize { get { return _key.KeySize; } }
-    }
-
-    public class FaultingKeyedHashAlgorithm : KeyedHashAlgorithm
-    {
-        KeyedHashAlgorithm _keyedHashAlgorithm;
-        Exception _throwMe;
-        byte[] _key;
-
-        public FaultingKeyedHashAlgorithm(KeyedHashAlgorithm keyedHashAlgorithm, Exception throwMe, byte[] key)
-        {
-            _keyedHashAlgorithm = keyedHashAlgorithm;
-            _throwMe = throwMe;
-            _key = key;
-        }
-
-        public override byte[] Key
-        {
-            get
-            {
-                if (_throwMe != null)
-                {
-                    throw _throwMe;
-                }
-
-                return _key;
-            }
-
-            set
-            {
-                _key = value;
-            }
-        }
-
-        protected override void HashCore(byte[] array, int ibStart, int cbSize)
-        {
-            throw new NotImplementedException();
-        }
-
-        protected override byte[] HashFinal()
-        {
-            throw new NotImplementedException();
-        }
-
-        public override void Initialize()
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    /// <summary>
-    /// Helpful for extensibility testing for errors.
-    /// </summary>
-    public class AlwaysReturnNullSignatureProviderFactory : SignatureProviderFactory
-    {
-        public override SignatureProvider CreateForSigning(SecurityKey key, string algorithm)
-        {
-            return null;
-        }
-
-        public override SignatureProvider CreateForVerifying(SecurityKey key, string algorithm)
-        {
-            return null;
-        }
-    }
 }
