+++ test/Microsoft.IdentityModel.Tokens.Jwt.Tests/JwtPayloadTest.cs
@@ -20,6 +20,7 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Security.Claims;
+using Microsoft.IdentityModel.Test;
 using Xunit;
 
 namespace Microsoft.IdentityModel.Tokens.Jwt.Test
