+++ test/Microsoft.IdentityModel.Tokens.Saml.Tests/project.json
@@ -1,23 +1,28 @@
 {
-    "version": "5.0.0-*",
-    "dependencies": {
-      "Newtonsoft.Json": "6.0.6",
-      "Microsoft.IdentityModel.Protocols": "2.0.0-*",
-      "Microsoft.IdentityModel.Tokens.Saml": "5.0.0-*",
-      "xunit.runner.aspnet": "2.0.0-aspnet-*"
-    },
-    "compile": [ "**\\*.cs", "..\\Shared\\**\\*.cs" ],
-    "commands": {
-        "test": "xunit.runner.aspnet"
-    },
-    "frameworks": {
-        "dnx451": {
-            "dependencies": {
-            }
-        }
-    },
-    "compilationOptions": {
-        "allowUnsafe": false,
-        "warningsAsErrors": true
+  "version": "5.0.0-*",
+  "dependencies": {
+    "Newtonsoft.Json": "6.0.6",
+    "Microsoft.IdentityModel.Protocols": "2.0.0-*",
+    "Microsoft.IdentityModel.Tokens.Saml": "5.0.0-*",
+    "xunit.runner.aspnet": "2.0.0-aspnet-*"
+  },
+  "exclude": [
+    "SamlSecurityTokenHandlerTests.cs",
+    "Saml2SecurityTokenHandlerTests.cs"
+  ],
+
+  "compile": [ "**\\*.cs", "..\\Shared\\**\\*.cs" ],
+  "commands": {
+    "test": "xunit.runner.aspnet"
+  },
+  "frameworks": {
+    "dnx451": {
+      "dependencies": {
+      }
     }
+  },
+  "compilationOptions": {
+    "allowUnsafe": false,
+    "warningsAsErrors": true
+  }
 }
