+++ test/Shared/ClaimSets.cs
@@ -20,6 +20,7 @@ using System;
 using System.Collections.Generic;
 using System.Security.Claims;
 using Microsoft.IdentityModel.Tokens;
+using Microsoft.IdentityModel.Tokens.Jwt;
 
 namespace Microsoft.IdentityModel.Test
 {
